chore: fixes lint issues

+++ pkg/pattern/drain/log_cluster.go
@@ -82,7 +82,7 @@ func (x *Volume) ForRange(start, end model.Time) *Volume {
 func (x *Volume) Matches() int64 {
 	var m int64
 	for i := range x.Values {
-		m += int64(x.Values[i].Value)
+		m += x.Values[i].Value
 	}
 	return m
 }
